# Improve performance of `fetch_all/4`